Turn templates into external files.
